+++ components/Button/index.js
@@ -1,4 +1,4 @@
-import Button from './button.vue';
+import Button from './button.js';
 
 Button.install = function (Vue) {
     Vue.component(Button.name, Button);