+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 ## 更新日志
+
+### New features
+  * Button
+      通过jsx的形式开发，实现链接按钮的【href】【target】属性，增加setButtonLink方法，设置button的href
+
 *2021-11-17*
 ## v1.0.28
  ## revert